fix.도메인 중심으로 프로젝트 구조 변경
