+++ src/main/java/com/ssginc/unnie/board/controller/BoardController.java
@@ -1,0 +1,4 @@
+package com.ssginc.unnie.board.controller;
+
+public class BoardController {
+}
