+++ src/main/java/com/ssginc/unnie/board/dto/BoardRequest.java
@@ -1,0 +1,4 @@
+package com.ssginc.unnie.board.dto;
+
+public class BoardRequest {
+}
