+++ src/main/java/com/ssginc/unnie/board/entity/Board.java
@@ -1,0 +1,4 @@
+package com.ssginc.unnie.board.entity;
+
+public class Board {
+}
