+++ src/main/java/com/ssginc/unnie/board/repository/BoardRepository.java
@@ -1,0 +1,4 @@
+package com.ssginc.unnie.board.repository;
+
+public class BoardRepository {
+}
