+++ src/main/java/com/ssginc/unnie/board/service/BoardService.java
@@ -1,0 +1,4 @@
+package com.ssginc.unnie.board.service;
+
+public interface BoardService {
+}
