+++ src/main/java/com/ssginc/unnie/board/service/impl/BoardServiceImpl.java
@@ -1,0 +1,4 @@
+package com.ssginc.unnie.board.service.impl;
+
+public class BoardServiceImpl {
+}
