+++ src/main/java/com/ssginc/unnie/common/exception/UnnieBoardException.java
@@ -1,0 +1,7 @@
+package com.ssginc.unnie.common.exception;
+
+public class UnnieBoardException extends UnnieCustomException {
+    public UnnieBoardException(String message) {
+        super(message);
+    }
+}
