--- src/main/java/com/ssginc/unnie/controller/BoardController.java
@@ -1,4 +1,0 @@
-package com.ssginc.unnie.controller;
-
-public class BoardController {
-}
