--- src/main/java/com/ssginc/unnie/dto/BoardResponse.java
@@ -1,4 +1,0 @@
-package com.ssginc.unnie.dto;
-
-public class BoardResponse {
-}
