--- src/main/java/com/ssginc/unnie/entity/Board.java
@@ -1,4 +1,0 @@
-package com.ssginc.unnie.entity;
-
-public class Board {
-}
