--- src/main/java/com/ssginc/unnie/exception/UnnieBoardException.java
@@ -1,7 +1,0 @@
-package com.ssginc.unnie.exception;
-
-public class UnnieBoardException extends UnnieCustomException {
-    public UnnieBoardException(String message) {
-        super(message);
-    }
-}
