--- src/main/java/com/ssginc/unnie/repository/BoardRepository.java
@@ -1,4 +1,0 @@
-package com.ssginc.unnie.repository;
-
-public class BoardRepository {
-}
