--- src/main/java/com/ssginc/unnie/service/BoardService.java
@@ -1,4 +1,0 @@
-package com.ssginc.unnie.service;
-
-public interface BoardService {
-}
