--- src/main/java/com/ssginc/unnie/service/impl/BoardServiceImpl.java
@@ -1,4 +1,0 @@
-package com.ssginc.unnie.service.impl;
-
-public class BoardServiceImpl {
-}
